+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2021-07-22
+
+### Added
+
+- Adds setting field **targetLocales** for the list of Locales that only need to be built.
+- Adds setting filed **excludedLocales** for the list of Locales that should be excluded in the build.
+
 ## [2.0.0] - 2021-06-30
 
 ### Added